// src/app/interfaces/course-details.ts
export interface CourseDetails {
  courseDetailsId?: number;
  courseCode: string;
  examType: string;
  courseType: string;
  classSchedule: string;
  internetAddress: string;
  coursePrice: number;
  minHeadcount: number;
  maxHeadcount: number;
  preliminaryRequirement: string;
  classPerWeek: number;
  classPerTerm: number;
  organizationName: string;
  languageName: string;
}